AI/requirements_enhanced.txt
# 核心依赖
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scikit-learn>=1.3.0
scipy>=1.11.0

# AI功能
openai>=1.0.0
requests>=2.31.0

# 数据处理
openpyxl>=3.1.0
xlrd>=2.0.1

# 文档处理
python-docx>=0.8.11
PyPDF2>=3.0.1

# 统计分析
statsmodels>=0.14.0
pingouin>=0.5.3

# 可视化增强
kaleido>=0.2.1
bokeh>=3.2.0

# 开发和测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 系统工具
psutil>=5.9.0
python-dotenv>=1.0.0

# 网络请求
httpx>=0.24.0
aiohttp>=3.8.0

# 数据验证
pydantic>=2.0.0

# 日志和配置
loguru>=0.7.0
pyyaml>=6.0